+++ CHANGELOG.md
@@ -35,5 +35,6 @@ Exceptions are acceptable depending on the circumstances (critical bug fixes tha
 
 ### Fixed
 
+- fixed json tags in requests and responses
 - fixed `golangci-lint` execution
 - fixed wrong image name in `workflow.yaml`